move hazard config to psha_adapter package
