--- nzshm_model/hazard_config.py
@@ -1,5 +1,0 @@
-from abc import ABC
-
-
-class HazardConfig(ABC):
-    pass
